Clean up utils
Remove unused methods and consolidate tests.
